feat(web): kleros-app
