+++ kleros-app/package.json
@@ -2,17 +2,16 @@
   "name": "@kleros/kleros-app",
   "version": "1.0.0",
   "source": "src/lib/index.ts",
-  "main": "dist/index.js",
-  "types": "dist/types.d.ts",
-  "isLibrary": true,
+  "main": "dist/main.js",
+  "module": "dist/module.js",
+  "types": "dist/index.d.ts",
   "scripts": {
-    "clear": "rm -r .parcel-cache",
+    "clear": "rm -r ../.parcel-cache",
     "clean": "rm -rf dist",
     "start": "parcel src/index.html",
-    "build": "parcel build",
+    "build": "yarn clear & yarn clean & yarn parcel build",
     "check-style": "eslint 'src/**/*.{ts,tsx}' -fix",
-    "check-types": "tsc --noEmit",
-    "watch": "parcel watch"
+    "check-types": "tsc --noEmit"
   },
   "repository": {
     "type": "git",
@@ -49,7 +48,16 @@
     "typescript": "^5.6.3"
   },
   "dependencies": {
+    "@kleros/ui-components-library": "^2.15.0",
+    "jose": "^5.9.6"
+  },
+  "peerDependencies": {
+    "@tanstack/react-query": "^5.59.20",
+    "graphql": "^16.9.0",
+    "graphql-request": "^7.1.2",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "viem": "^2.21.42",
+    "wagmi": "^2.12.27"
   }
 }
