+++ kleros-app/src/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>My First Parcel App</title>
+    <title>Kleros App</title>
   </head>
   <body>
     <div id="app"></div>
