+++ kleros-app/src/utils/index.ts
@@ -1,0 +1,2 @@
+export const isUndefined = (maybeObject: any): maybeObject is undefined | null =>
+  typeof maybeObject === "undefined" || maybeObject === null;
