+++ web/package.json
@@ -74,6 +74,7 @@
   },
   "dependencies": {
     "@cyntler/react-doc-viewer": "^1.16.3",
+    "@kleros/kleros-app": "workspace:^",
     "@kleros/kleros-sdk": "workspace:^",
     "@kleros/kleros-v2-contracts": "workspace:^",
     "@kleros/ui-components-library": "^2.15.0",
